USB: serial: cp210x: Add GPIO support for CP2104.
This patch adds support for GPIO on CP2104.

I successfully tested input and output in open-drain mode (which is the
default, unless push-pull is configured in on-chip EEPROM).

Probably some of new functions may be reused for GPIO on CP2103, but I
do not have any CP2103-based devices for testing.

Signed-off-by: Ivan Mironov <mironov.ivan@gmail.com>